Refactor things out of __init__.py in a lot of cases.
Because classes in files makes sense.
