+++ trafficlight/store.py
@@ -15,7 +15,7 @@
 import logging
 from typing import List, Optional
 
-from trafficlight.objects import Client
+from trafficlight.objects.client import Client
 from trafficlight.tests import TestCase, TestSuite
 
 logger = logging.getLogger(__name__)
