+++ trafficlight/tests/__init__.py
@@ -24,7 +24,8 @@ from typing import Callable, List, Optional
 import trafficlight.homerunner
 from trafficlight.client_types import ClientType
 from trafficlight.homerunner import HomeserverConfig
-from trafficlight.objects import Client, Model
+from trafficlight.objects.client import Client
+from trafficlight.objects.model import Model, ModelState
 from trafficlight.server_types import ServerType
 from trafficlight.tests.assertions import TestException
