+++ trafficlight/tests/assertions.py
@@ -14,8 +14,7 @@
 # limitations under the License.
 from typing import Any
 
-from trafficlight.objects import Model
-
+from trafficlight.objects.model import Model
 #
 # Aware that we're skipping naming conventions here.
 # Doing so to keep in sync with unittest -style (junit style?) conventions
