+++ trafficlight/tests/send_messages_testsuite.py
@@ -3,7 +3,8 @@ from typing import List
 
 import trafficlight.tests
 from trafficlight.homerunner import HomeserverConfig
-from trafficlight.objects import Client, Model, ModelState
+from trafficlight.objects.client import Client
+from trafficlight.objects.model import Model, ModelState
 from trafficlight.tests.assertions import assertCompleted
